Add code docs for RPC server functions
The test for RpcServer.Close() was returning too soon. The simplest
solution is an ugly sleep of 1 second.
